projects/active/requirements.txt
aiohttp>=3.8.0
asyncio
psutil>=5.9.0
jsonschema>=4.0.0
pyyaml>=6.0
watchdog>=3.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
redis[hiredis]>=4.5.0
hiredis>=2.2.0
pyjwt>=2.8.0
bcrypt>=4.0.0
cryptography>=41.0.0
passlib>=1.7.4